requirements.txt
pandas
numpy
matplotlib
yfinance
python-dotenv
openai
